refactor: reducer 파일 분리
